src/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
pydantic-settings==2.12.0
yt-dlp==2025.11.12
motor==3.7.1
faster-whisper==1.2.1
langchain-text-splitters==1.0.0
langchain-community==0.4.1
qdrant-client==1.16.0
sentence-transformers==5.1.2

